adjust example service and docker build file (#45)
* adjust example service and docker build file
fix exception handling in keystone.py
improve documentation

* make lint happy

* removes notes

+++ denbi/scripts/perun_propagation_service.py
@@ -36,49 +36,52 @@ from denbi.perun.keystone import KeyStone
 from flask import Flask
 from flask import request
 
-
-app = Flask(__name__)
-app.config['cleanup'] = True
-app.config['keystone_read_only'] = os.environ.get('KEYSTONE_READ_ONLY', 'False').lower() == 'true'
-
-#  configure basic logger
-logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s -%(message)s')
+import traceback
 
 # logging formatter
-fmt = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s -%(message)s')
-
-# create a FileHandler for reporting
-# report_ch = logging.FileHandler("report.log")
+fmt = logging.Formatter('[%(asctime)s] - (%(name)s/%(levelname)s) - %(message)s')
 
 # create a StreamHandler for reporting
 report_ch = logging.StreamHandler(sys.stdout)
 report_ch.setLevel(logging.INFO)
 report_ch.setFormatter(fmt)
 
-# configure logger for report_domain using default name
+# configure 'report' logger
 report = logging.getLogger('report')
 report.setLevel(logging.INFO)
 report.addHandler(report_ch)
 
+app = Flask(__name__)
+
+# Load configuration from file
+if "CONFIG_PATH" not in os.environ:
+    os.environ['CONFIG_PATH'] = f"{os.getcwd()}/perun_propagation_service.cfg"
+
+if os.path.exists(os.environ['CONFIG_PATH']):
+    report.info(f"Loading configuration from {os.environ['CONFIG_PATH']}.")
+    app.config.from_pyfile(os.environ['CONFIG_PATH'])
+else:
+    report.info(f"Configuration file {os.environ['CONFIG_PATH']} not found. Using defaults.")
+
 # create a FileHandler for logging
-log_ch = logging.FileHandler("pka.log")
-log_ch.setLevel(logging.ERROR)
+log_ch = logging.FileHandler(app.config.get("LOG_DIR", "") + "/pka.log")
+log_ch.setLevel(logging.INFO)
 log_ch.setFormatter(fmt)
 
-# configure logger for logger_domain using default name
+# configure 'denbi' logger
 denbi = logging.getLogger('denbi')
-denbi.setLevel(logging.ERROR)
+denbi.setLevel(logging.INFO)
 denbi.addHandler(log_ch)
 
-
+# Create thread executor
 executor = ThreadPoolExecutor(max_workers=1)
 
 
 def process_tarball(tarball_path, base_dir=tempfile.mkdtemp(), read_only=False, target_domain_name='elixir',
-                    default_role='user', nested=False, support_quota=False, cloud_admin=True):
+                    default_role='user', nested=False, support_quota=False, cloud_admin=True, cleanup=False):
     """Process Perun tarball."""
     d = datetime.today()
-    dir = base_dir + "/" + str(d.year) + "_" + str(d.month) + "_" + str(d.day) + "-" + str(d.hour) + ":" + str(d.minute) + ":" + str(d.second) + "." + str(d.microsecond)
+    dir = f"{base_dir}/{d.year}_{d.month}_{d.day}_{d.hour}:{d.minute}:{d.second}.{d.microsecond}"
     os.mkdir(dir)
 
     report.info("Processing data uploaded by Perun: %s" % tarball_path)
@@ -95,16 +98,18 @@ def process_tarball(tarball_path, base_dir=tempfile.mkdtemp(), read_only=False, 
 
     endpoint = Endpoint(keystone=keystone, mode="denbi_portal_compute_center",
                         support_quotas=support_quota)
+
     endpoint.import_data(dir + '/users.scim', dir + '/groups.scim')
     report.info("Finished processing %s" % tarball_path)
 
     # Cleanup
-    shutil.rmtree(dir)
+    if cleanup:
+        shutil.rmtree(dir)
 
 
 @app.route("/upload", methods=['PUT'])
 def upload():
-    """Recieve a perun tarball, store it in a temporary file and process it."""
+    """Receive a perun tarball, store it in a temporary file and process it."""
 
     # Create a tempfile to write the data to. delete=False because we will
     # close after writing, before processing, and this would normally cause a
@@ -115,14 +120,22 @@ def upload():
     file.write(request.get_data())
     file.close()
 
-    # execute
-    executor.submit(process_tarball, file.name, read_only=app.config.get('KEYSTONE_READ_ONLY', False),
-                    target_domain_name=app.config.get('TARGET_DOMAIN_NAME', 'elixir'),
-                    default_role=app.config.get('DEFAULT_ROLE', 'user'),
-                    nested=app.config.get('NESTED', False),
-                    cloud_admin=app.config.get('CLOUD_ADMIN', True),
-                    base_dir=app.config.get('BASE_DIR', tempfile.mkdtemp()),
-                    support_quota=app.config.get('SUPPORT_QUOTA', False))
+    # execute task
+    result = executor.submit(process_tarball, file.name, read_only=app.config.get('KEYSTONE_READ_ONLY', False),
+                             target_domain_name=app.config.get('TARGET_DOMAIN_NAME', 'elixir'),
+                             default_role=app.config.get('DEFAULT_ROLE', 'user'),
+                             nested=app.config.get('NESTED', False),
+                             cloud_admin=app.config.get('CLOUD_ADMIN', True),
+                             base_dir=app.config.get('BASE_DIR', tempfile.mkdtemp()),
+                             support_quota=app.config.get('SUPPORT_QUOTA', False),
+                             cleanup=app.config.get('CLEANUP', False))
+
+    # if task fails with an exception, the thread pool catches the exception,
+    # stores it, then re-raises it when we call the result() function.
+    try:
+        result.result()
+    except Exception:
+        traceback.print_exc()
 
     if app.config.get('CLEANUP', False):
         os.unlink(file)
@@ -130,6 +143,5 @@ def upload():
     return ""
 
 
-app.config.from_envvar('CONFIG_PATH')
 if __name__ == "__main__":
     app.run(host=app.config['HOST'], port=app.config['PORT'])
